P1-9 terminado

+++ EjerciciosU1/ejercicio11.py
@@ -1,3 +1,3 @@
-n= float(input("Introduce un número positivo: "))
+n= int(input("Introduce un número positivo: "))
 suma = n * (n + 1)/2
 print("suma =", suma)
